ignoring root alias
